pre-aplha of the app!
I think the app is developed enough to enter a pre-alpha phase! There are still many things that need to be fixed or added, but the main idea should work! Things to be added: adding files to root folder, remembering the root folder, adding folder structure to the FileList.jsx, adding folders to root folder, remove the console.logs in main for debug as well as refine the comments while also removing the handles that won't be needed anymore. Things that are done: modifying file / reading files, getting a directory with all the files, working text editor.